requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.44.0
cmake==3.31.0.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
crispy-bootstrap4==2024.10
cryptography==43.0.3
cssselect==1.2.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-heroku==0.3.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
fqdn==1.5.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector-python==9.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pillow==10.2.0
pipenv==2024.4.1
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
referencing==0.35.1
reportlab==4.4.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
Scrapy==2.11.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tldextract==5.1.2
tornado==6.4.2
traitlets==5.14.3
Twisted==24.10.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.0.7
virtualenv==20.29.2
w3lib==2.2.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.6.0
zope.interface==7.1.1
